remove packages that were moved to common, create hierarchy and publish to PYPI
