+++ README.md
@@ -1,1 +1,3 @@
 # BallBoi
+
+BallBoi - A package to fetch sports information from https://profootballapi.com
