+++ strotools.ballboi.egg-info/SOURCES.txt
@@ -1,0 +1,10 @@
+README.md
+setup.py
+strotools.ballboi.egg-info/PKG-INFO
+strotools.ballboi.egg-info/SOURCES.txt
+strotools.ballboi.egg-info/dependency_links.txt
+strotools.ballboi.egg-info/requires.txt
+strotools.ballboi.egg-info/top_level.txt
+strotools/ballboi/__init__.py
+strotools/ballboi/nfl_retriever.py
+strotools/ballboi/retriever.py
