+++ strotools.ballboi.egg-info/requires.txt
@@ -1,0 +1,1 @@
+strotools.common
