+++ strotools/ballboi/nfl_retriever.py
@@ -8,13 +8,13 @@ import logging
 #------------------------------------------------------------------------------
 # Non-Standard Library From Imports
 #------------------------------------------------------------------------------
-from connection import Connection
-from retriever  import RetrieverInterface
-from util       import build_params
+from strotools.ballboi.retriever  import RetrieverInterface
+from strotools.common.connection  import Connection
+from strotools.common.util        import build_params
 #------------------------------------------------------------------------------
 # Exceptions
 #------------------------------------------------------------------------------
-from util import InvalidParamsError
+from strotools.common.util import InvalidParamsError
 
 class InvalidSeasonTypeError(Exception):
     ''' Thrown if incorrect season type is provided '''
